requirements.txt
pytest
pytest-playwright
pytest-html
requests
pydantic
python-dotenv
allure-pytest